+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### New in 0.16.0 (Released 2018/10/05)
+
+- [__#39__](https://github.com/spectresystems/spectre.cli/issues/39) No suggestions for root command
+- [__#38__](https://github.com/spectresystems/spectre.cli/issues/38) Default command not registered with type registrar
+
 ### New in 0.15.0 (Released 2018/10/04)
 
 - [__#34__](https://github.com/spectresystems/spectre.cli/issues/34) Allow relaxed option parsing